Windows port

+++ lib/filter/string.cpp
@@ -203,7 +203,7 @@ int filter_string_date_seconds_since_epoch ()
 {
   auto now = chrono::system_clock::now ();
   auto duration = now.time_since_epoch ();
-  int seconds = chrono::duration_cast<std::chrono::seconds>(duration).count();
+  int seconds = (int) chrono::duration_cast<std::chrono::seconds>(duration).count();
   return seconds;
 }
